Alarms are now shown depending on current time
Alarms are now seperated depending on the current time.

Recycler view has a logic error of displaying an empty view if a drug
added to database and does not show for the current time period. Will
fix this when I wake up

If database is empty, a message prompts users to add new reminder

To do:
notifications for each item in database
fix recycler view to not show empty views